+++ runtime/test/unit/runner.clj
@@ -22,7 +22,7 @@
             [ujima.linux.audio-test]
             [ujima.linux.i3-test]
             [ujima.linux.systemd-test]
-            [ujima.desktop.http-test]
+            [lib.http-test]
             [ujima.desktop.http.ui-test]
             [ujima.desktop.app.catalog-test]
             [ujima.desktop.app-test]
@@ -54,7 +54,7 @@
     ujima.linux.audio-test
     ujima.linux.i3-test
     ujima.linux.systemd-test
-    ujima.desktop.http-test
+    lib.http-test
     ujima.desktop.http.ui-test
     ujima.desktop.app.catalog-test
     ujima.desktop.app-test